tests/system_tests/requirements.txt
# System test dependencies
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-xdist>=3.0.0
pyyaml>=6.0
requests>=2.28.0
selenium>=4.0.0
webdriver-manager>=3.8.0
